uievents/legacy-domevents-tests/approved/dispatchEvent.click.checkbox.html

<!DOCTYPE html>
<html>
<head>
<title> MouseEvent: Default action and synthetic click event </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>

<div style="">
    <input type="checkbox" id="target">
    <button id="button"> Click Here </button>
</div>

<script>
async_test(function() {
    var EVENT = "click";
    var TARGET = document.getElementById("target");
    var BUTTON = document.getElementById("button");
    var state;

    var TestEvent = this.step_func(function(evt) {
        if (BUTTON == evt.target) {
            BUTTON.removeEventListener(EVENT, TestEvent, true);

            var e = document.createEvent("MouseEvent");
            e.initMouseEvent(EVENT,  /* type */
                             false,  /* bubbles */
                             true,   /* cancelable */
                             window, /* view */
                             1,      /* detail */
                             0,      /* screenX */
                             0,      /* screenY */
                             0,      /* clientX */
                             0,      /* clientY */
                             false,  /* ctrlKey */
                             false,  /* altKey */
                             false,  /* shiftKey */
                             false,  /* metaKey */
                             0,      /* button */
                             null    /* relatedTarget */);

            assert_equals(TARGET.checked, state,
                          "Checkbox state is unchanged before the synthetic click event is dispatched");
            assert_equals(e.type, EVENT, "e.type");
            assert_false(e.bubbles, "e.bubbles");

            TARGET.dispatchEvent(e);
        } else if (TARGET == evt.target) {
            assert_array_equals(TARGET.checked, !state,
                                "Default action is performed when a synthetic click event is dispatched on a checkbox element");
            assert_equals(TARGET.checked, state,
                          "Checkbox state is unchanged before the synthetic click event is dispatched");
            assert_equals(evt.type, EVENT, "evt.type");
            assert_false(evt.bubbles, "evt.bubbles");


            this.done();
        }
    });

    BUTTON.addEventListener(EVENT, TestEvent, true);
    TARGET.addEventListener(EVENT, TestEvent, true);

    window.onload = this.step_func(function() {
        state = TARGET.checked;
        BUTTON.click();
    });
});
</script>
</body>
</html>
